pyproject.toml
[project]
name = "learn-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain-community (>=0.3.18,<0.4.0)",
    "langchain-core>=0.3.40,<0.4.0",
    "langchain-openai>=0.3.7",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-google-genai (>=2.0.11,<3.0.0)",
    "crewai",
    "crewai-tools",
    "langgraph>=0.3.2",
    "typing-extensions>=4.12.2",
    "ipykernel>=6.29.5",
]
